pyproject.toml
[project]
name = "web-collect-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioconsole>=0.8.2",
    "loguru>=0.7.3",
    "patchright>=1.55.2",
    "ruamel-yaml>=0.18.16",
]
